rm: connection manager
