+++ duva/src/actor_registry.rs
@@ -1,6 +1,5 @@
 use crate::domains::cluster_actors::commands::ClusterCommand;
 use crate::domains::config_actors::config_manager::ConfigManager;
-use crate::presentation::clusters::connection_manager::ClusterConnectionManager;
 
 use crate::{
     CacheManager, presentation::clusters::communication_manager::ClusterCommunicationManager,
@@ -17,9 +16,5 @@ pub struct ActorRegistry {
 impl ActorRegistry {
     pub(crate) fn cluster_communication_manager(&self) -> ClusterCommunicationManager {
         ClusterCommunicationManager(self.cluster_actor_handler.clone())
-    }
-
-    pub(crate) fn cluster_connection_manager(&self) -> ClusterConnectionManager {
-        ClusterConnectionManager(self.cluster_communication_manager())
     }
 }
