--- duva/src/presentation/clusters/connection_manager.rs
@@ -1,85 +1,0 @@
-use super::communication_manager::ClusterCommunicationManager;
-use super::outbound::stream::OutboundStream;
-use crate::domains::cluster_actors::commands::ClusterCommand;
-use crate::domains::peers::identifier::PeerIdentifier;
-use crate::{InboundStream, make_smart_pointer};
-use std::collections::VecDeque;
-use tokio::sync::mpsc::Sender;
-
-pub(crate) struct ClusterConnectionManager(pub(crate) ClusterCommunicationManager);
-
-impl ClusterConnectionManager {
-    pub(crate) async fn accept_inbound_stream(
-        &self,
-        mut peer_stream: InboundStream,
-        ccm: ClusterCommunicationManager,
-    ) -> anyhow::Result<()> {
-        let connected_peer_info = peer_stream.recv_handshake().await?;
-
-        peer_stream.disseminate_peers(self.0.get_peers().await?).await?;
-
-        let (callback, rx) = tokio::sync::oneshot::channel();
-        let add_peer_cmd =
-            peer_stream.prepare_add_peer_cmd(ccm, connected_peer_info, callback).await?;
-        self.send(add_peer_cmd).await?;
-        rx.await?;
-        Ok(())
-    }
-
-    pub(crate) async fn discover_cluster(
-        self,
-        self_port: u16,
-        connect_to: PeerIdentifier,
-    ) -> anyhow::Result<()> {
-        let mut queue = VecDeque::from(vec![connect_to.clone()]);
-        let mut callbacks = Vec::new();
-        while let Some(connect_to) = queue.pop_front() {
-            let (tx, callback) = tokio::sync::oneshot::channel();
-            queue.extend(
-                ClusterConnectionManager(self.clone())
-                    .connect_to_server(self_port, connect_to, tx)
-                    .await?,
-            );
-            callbacks.push(callback);
-        }
-        for callback in callbacks {
-            let _ = callback.await;
-        }
-        Ok(())
-    }
-
-    async fn connect_to_server(
-        self,
-        self_port: u16,
-        connect_to: PeerIdentifier,
-        callback: tokio::sync::oneshot::Sender<()>,
-    ) -> anyhow::Result<Vec<PeerIdentifier>> {
-        let existing_peers = self.get_peers().await?;
-        if existing_peers.contains(&connect_to) {
-            return Ok(vec![]);
-        }
-
-        let replication_info = self.replication_info().await?;
-        let stream = OutboundStream::new(connect_to, replication_info)
-            .await?
-            .initiate_handshake(self_port)
-            .await?
-            .set_replication_info(&self)
-            .await?;
-
-        let (add_peer_cmd, peer_list) = stream.create_peer_cmd(self.handler())?;
-        self.send(ClusterCommand::AddPeer(add_peer_cmd, callback)).await?;
-
-        Ok(peer_list)
-    }
-
-    pub async fn send(&self, cmd: ClusterCommand) -> anyhow::Result<()> {
-        Ok(self.0.send(cmd).await?)
-    }
-
-    fn handler(&self) -> Sender<ClusterCommand> {
-        self.0.0.clone()
-    }
-}
-
-make_smart_pointer!(ClusterConnectionManager, ClusterCommunicationManager);
