+++ duva/src/presentation/clusters/mod.rs
@@ -1,5 +1,5 @@
 pub mod communication_manager;
-pub mod connection_manager;
+
 pub mod inbound;
 pub mod listeners;
 pub mod outbound;
